requirements.txt
numpy==1.26.4
matplotlib==3.10.1
tqdm==4.67.1
numba==0.61.2
PyYAML==6.0.2